Predict stripping
Stops mob verbs from getting moved around again.
