+++ Content.Shared/Ensnaring/SharedEnsnareableSystem.cs
@@ -1,6 +1,8 @@
 using Content.Shared.DoAfter;
 using Content.Shared.Ensnaring.Components;
+using Content.Shared.IdentityManagement;
 using Content.Shared.Movement.Systems;
+using Content.Shared.Popups;
 using Robust.Shared.GameStates;
 using Robust.Shared.Serialization;
 
@@ -15,6 +17,8 @@ public abstract class SharedEnsnareableSystem : EntitySystem
 {
     [Dependency] private readonly MovementSpeedModifierSystem _speedModifier = default!;
     [Dependency] protected readonly SharedAppearanceSystem Appearance = default!;
+    [Dependency] protected readonly SharedPopupSystem Popup = default!;
+    [Dependency] private readonly SharedDoAfterSystem _doAfter = default!;
 
     public override void Initialize()
     {
@@ -84,5 +88,38 @@ public abstract class SharedEnsnareableSystem : EntitySystem
             return;
 
         args.ModifySpeed(component.WalkSpeed, component.SprintSpeed);
+    }
+
+    /// <summary>
+    /// Used where you want to try to free an entity with the <see cref="EnsnareableComponent"/>
+    /// </summary>
+    /// <param name="target">The entity that will be freed</param>
+    /// <param name="user">The entity that is freeing the target</param>
+    /// <param name="ensnare">The entity used to ensnare</param>
+    /// <param name="component">The ensnaring component</param>
+    public void TryFree(EntityUid target, EntityUid user, EntityUid ensnare, EnsnaringComponent component)
+    {
+        // Don't do anything if they don't have the ensnareable component.
+        if (!HasComp<EnsnareableComponent>(target))
+            return;
+
+        var freeTime = user == target ? component.BreakoutTime : component.FreeTime;
+        var breakOnMove = !component.CanMoveBreakout;
+
+        var doAfterEventArgs = new DoAfterArgs(EntityManager, user, freeTime, new EnsnareableDoAfterEvent(), target, target: target, used: ensnare)
+        {
+            BreakOnMove = breakOnMove,
+            BreakOnDamage = false,
+            NeedHand = true,
+            BreakOnDropItem = false,
+        };
+
+        if (!_doAfter.TryStartDoAfter(doAfterEventArgs))
+            return;
+
+        if (user == target)
+            Popup.PopupPredicted(Loc.GetString("ensnare-component-try-free", ("ensnare", ensnare)), target, target);
+        else
+            Popup.PopupPredicted(Loc.GetString("ensnare-component-try-free-other", ("ensnare", ensnare), ("user", Identity.Entity(target, EntityManager))), user, user);
     }
 }
